Prevent image to go out of screen
Limit the image position to the screen, so it cannot disappear from the
display.
While ooing out, the image is centered, if it is small enough to fit the
screen. As currently no relative information is stored, zooming does not
retain the centered status, thus zooming in makes the image uncentered.